xterm: Move fully to xterm, deprecate urxvt, follow color mode in vim
That way, key combos like ^-Shift-W work in the terminal again (urxvt doesn't
discern ^W and ^-Shift-W). Track the dark mode so that vim chooses the right
mode, too. Ideally, I'd have this as a global dynamic setting, but this is Unix,
not a proper desktop operating system like macOS or Windows.